requirements.txt
absl-py==2.1.0
altair==5.4.1
altgraph==0.17.4
attrs==24.2.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
docutils==0.21.2
ffmpeg==1.4
flatbuffers==24.3.25
fonttools==4.54.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
jax==0.4.35
jaxlib==0.4.35
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.0
mutagen==1.47.0
narwhals==1.13.3
numpy==2.1.3
opencv-contrib-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.0.0
protobuf==5.28.3
psutil==6.1.0
pyarrow==18.0.0
pycparser==2.22
pycryptodomex==3.21.0
pydeck==0.9.1
pyee==12.0.0
Pygments==2.18.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pyparsing==3.2.0
pypiwin32==223
pyshorteners==1.0.1
python-dateutil==2.9.0.post0
python-ffmpeg==2.0.12
pytube==15.0.0
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
qrcode==8.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
scipy==1.14.1
setuptools==75.3.0
six==1.16.0
smmap==5.0.1
sounddevice==0.5.1
streamlit==1.40.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
websockets==13.1
youtube-dl==2021.12.17
youtube-transcript-api==0.6.2
yt-dlp==2024.11.4